Big commit * SafeLeave * PlayerReturner, TeleportingPlayerReturner * Improved PlayerFreezer * LocationParser

+++ src/main/java/me/supcheg/advancedmanhunt/timer/CountDownTimerBuilder.java
@@ -22,7 +22,7 @@ public interface CountDownTimerBuilder {
 
     @NotNull
     @Contract("_ -> this")
-    CountDownTimerBuilder times(int times);
+    CountDownTimerBuilder times(long times);
 
     @NotNull
     @Contract("_ -> this")
